# backend/docs/swagger.yaml
basePath: /api
definitions:
...
  handlers.CreateCategoryRequest:
    properties:
      description:
        type: string
      name:
        type: string
      order:
        type: integer
    required:
    - name
    type: object
  handlers.CreateMenuItemRequest:
    properties:
      allergens:
        type: string
      calories:
        type: integer
      carbs:
        type: number
      categoryId:
        type: string
      description:
        type: string
      fat:
        type: number
      fiber:
        type: number
      image:
        type: string
      name:
        type: string
      preparationTime:
        type: integer
      price:
        type: number
      protein:
        type: number
      sodium:
        type: number
    required:
    - categoryId
    - name
    - price
    type: object
...
  models.ErrorResponse:
    properties:
      error:
        example: Error message
        type: string
      message:
        example: Error occurred
        type: string
      success:
        example: false
        type: boolean
    type: object
...
  models.SuccessResponse:
    properties:
      data: {}
      message:
        example: Operation completed successfully
        type: string
      success:
        example: true
        type: boolean
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: A comprehensive food delivery application API
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Restaurant App API
  version: "1.0"
paths:
...
  /menu-items/{id}:
    get:
      consumes:
      - application/json
      description: Get detailed information about a specific menu item
      parameters:
      - description: Menu Item ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Get menu item details
      tags:
      - menu
  /menu/categories:
    post:
      consumes:
      - application/json
      description: Create a new menu category for restaurant
      parameters:
      - description: Category data
        in: body
        name: category
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateCategoryRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - Bearer: []
      summary: Create menu category
      tags:
      - menu
  /menu/items:
    post:
      consumes:
      - application/json
      description: Create a new menu item for restaurant
      parameters:
      - description: Menu item data
        in: body
        name: item
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateMenuItemRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - Bearer: []
      summary: Create menu item
      tags:
      - menu
  /menu/items/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a menu item
      parameters:
      - description: Menu Item ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - Bearer: []
      summary: Delete menu item
      tags:
      - menu
    put:
      consumes:
      - application/json
      description: Update menu item details
      parameters:
      - description: Menu Item ID
        in: path
        name: id
        required: true
        type: string
      - description: Menu item update data
        in: body
        name: item
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateMenuItemRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - Bearer: []
      summary: Update menu item
      tags:
      - menu
  /menu/items/{id}/toggle:
    patch:
      consumes:
      - application/json
      description: Toggle availability of a menu item
      parameters:
      - description: Menu Item ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - Bearer: []
      summary: Toggle menu item availability
      tags:
      - menu
...
  /public/restaurants/{id}/menu:
    get:
      consumes:
      - application/json
      description: Get complete menu with categories and items for a restaurant
      parameters:
      - description: Restaurant ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Get restaurant menu
      tags:
      - menu
...
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"